Add files via upload
Here is the ETL I coded it extracts data from a book site called as Idefix It sends HTTP requests, retrieves the HTML response via Requests library, and then uses BeautifulSoup library to parse the HTML and extract relevant information such as book names, publishers, and prices. Afterwards the data is transformed by removing HTML elememnts, unnecessary characters or whitespace, and finally the data is inserted to a MySQL database with the MySqlConnecter.py which uses mysql.connecter library.