Updated Files
Updated Book_Scraping.py so It collects more information about book such as : authors , publish date etc. Did it by collecting href elements of the books in site which has the book ID. However it came with a cost of speed. Now the code is much slower because it makes 2 separate requests to the site and collect more data. I will learn Async.io, io.http and multiprocessing in the future to make the code faster.